requirements.txt
weaviate-client 
langchain==0.2.11 
langchain-community==0.2.10
langchain-weaviate 
langchain_cohere 
cohere 
sentence-transformers
openai-whisper  
gradio
streamlit
fastapi
pydantic